Add: advanced settings UI (Derin Uyku + Pil Bildirim + Sessiz Mod)
- activity_main.xml: Switch + Spinner (123 → 227 lines)
- MainActivity.kt: listeners + initial values + imports
- Settings.kt: pilBildirimAraligiDk dedup
- EdaForegroundService.kt: Int*Long fix
- onStartCommand: ACTION_BASLAT + ACTION_DERIN_UYKU

+++ android-app/app/src/main/java/com/deniz/eda/service/EdaForegroundService.kt
@@ -151,7 +151,7 @@ class EdaForegroundService : Service(), TextToSpeech.OnInitListener {
         pilJob?.cancel()
         pilJob = serviceScope.launch {
             while (isActive) {
-                delay(Settings.pilBildirimAraligiDk * 60_000L)
+                delay(Settings.pilBildirimAraligiDk.toLong() * 60_000L)
                 val bilgi = BatteryUtils.pilBilgisiAl(this@EdaForegroundService)
                 if (bilgi != null) {
                     konus("Pil yüzde ${bilgi.yuzde} ${Settings.kullaniciAdi}.")
